Convert vectors in classic h2dust calc to scalars
Specifically, I converted the variables, d_indixe, d_t1, d_t2, d_tdef
dusti1, dusti2.

I get that the goal was to try to mirror some of the variables in
`logTlininterp_buf`, but there is currently no reason to allocate the
extra memory. Plus, this is sort of a one-off event.

If we really want to, we can make these into arrays again in the future