Introduce initial support for clang-tidy
For the uninitiated, clang-tidy is a highly customizable general-purpose
(and extendable) C/C++ linter.

Documentation And Usage
=======================

It's **really** unfortunate that we can't integrate this check with the
pre-commit framework. That means that locally installing and using the
tool is fairly manual (I tried to provide lots of details about this).

I have written a **LOT** of documentation for using this tool. This
involved reworking a bunch of existing docs about the formatters and
other checks that are run by pre-commit. But, hopefully this all makes
``clang-tidy`` straight-forward for anybody to understand). I also have
written some custom CMake logic to make things a lot easier to use.

Honestly, this was all substantially more work than I originally
anticipated. I probably would have said that I would not do it all over
again given the choice (I think it's valuable now that we have it, just
not a good use of my time).

With that said, just before making this commit I briefly tested some of
the (currently disabled) `clang-analyzer-` class of checks, and it looks
like they may have flagged some cases where we use unitialized variables.
I haven't dug into this very much, but this *could* make all of the
effort *very* worthwhile.

Current Checks
==============

Checks are **ONLY** run for C++ files.

For this initial commit, I've cast a fairly wide net on introducing
lints. I wasn't especially thorough with all of the available options
and I was pretty liberal with what I enabled (my primary criterion was
does it flag existing code).

My current philosophy is that we can continue disabling overly
burdensome checks as we continue porting code to C++. Once we're done,
we could re-enable a bunch of checks. (I could probably be convinced to
initially adopt a more conservative approach, but I figured that using
more strict checks for our unit-testing code would be good).

I have explicitly disabled:
- a bunch of stylistic checks. Aside: you could get very detailed about
  enforcing naming conventions (e.g. snake_case, SCREAMING_SNAKE_CASE,
  CamelCase, etc.) for various types of constructs. (You could spend
  endless time bikeshedding here)
- I also haven't enabled the `clang-analyzer-` family of checks. I did a
  quick check and this flags a bunch. Some of these checks (pertaining
  to unitialized variables) seem like **very useful** warnings that
  merit more detailed examination

Failing Checks
==============

Some of the checks should trigger some easy-to-fix (mostly stylistic)
failing checks in the current code-base. I've already resolved all of
these failure, but have explicitly excluded those fixes from this commit
so that we can confirm CI actually catches the errors.