Modify InternalGrUnits to control mh with enum
Previously, InternalGrUnits only needed to swap between 2 values so we
used a switch of 0 or 1. But, now we have added support for a 3rd value
of the Hydrogen Mass (i.e. the value historically used to compute units
for the cloudy table). Now that we have 3 values, we define an enum.

Don't forget, this is all temporary. The time is rapidly approaching
when we will be to transition to just having a single definition for the
hydrogen mass constant.