Introduce ``InjectPathFieldPack`` 1/3
Essentially, this type holds pointers to the various injection pathway
fields in the order that is implicitly assumed throughout the codebase.
This change helps simplify make_consistent and it will also simplify
`calc_grain_size_increment_1d` (once transcribed). Longer term, I have a
plan that will let use dispose of this ``InjectPathFieldPack`` type
(but, we need to make a few changes before we get to that).